refactor `new_community` to reduce complexity to O(1) using geohash buckets (#55)
This PR replaces https://github.com/encointer/pallets/pull/50.
closes #9 

* organize meetup location storage in buckets, addressed by geohash
* reduce minimal distance checks to adjacent buckets

It incorporates all changes discussed in the above PR and adds one more feature:
https://github.com/encointer/pallets/commit/b61b39617e157cfde9026c3bec3e60d225953d77 incorporates the datastructure changes from `communities` (due to geohash) into `ceremonies`. The main problem here was that we can no longer retrieve a simple list of all locations from a community, because there is a layer for geohashes in between. I solved the problem as follows:

1.  Add a function `pub fn get_locations(cid:&CommunityIdentifier) -> Vec<Location>` that uses `iter_prefix_values` on the `Locations` double map in simply concatenates all the vectors corresponding to the geohashes of one community.
2. Because we rely on the indices of the elements in the above retrieved vector in allocation and issuance, I prevented locations from being added/removed in phases other that `CeremonyPhaseType::REGISTERING`.

also, what is crucial for this to work is that `iter_prefix_values` reliably returns its elements in the same order (if the underlying datastructure does not change). Can we rely on that?

note that those changes are somewhat transitional, as the algorithms for issuance and assignment will change again. but i wanted to have a version that works with the current `communities` pallet in order for this PR to be merged.