[Ceremonies] fix potential runtime panic for ReputationLifetimes > 2 (#112)
An unsigned integer underrun caused an erroneous `ProofOutdated` error when trying to register with a proof of attendance on the app side. (because we increased the reputation lifetime to 337 recently).

Interestingly, the code panics when this happens in the unit tests, whereas nothing was observed when running the node, except for the above `ProofOutdated` error.